+++ becok/src/styles/CardModal/ContestModal.style.js
